refactor(docs): #704 PR-C — extract upload submodule
Second extraction in the routes/__init__.py shrink chain (#704). Moves
the multipart upload form (`_upload_form`, `_doc_type_radio`,
`_vtk_picker`, `_version_picker`), client-side JS payloads
(`_FILE_PICKER_SCRIPT`, `_DOC_TYPE_TOGGLE_SCRIPT`,
`_UPLOAD_MAX_BYTES`), upfront `parent_vtk_id` validation
(`_validate_parent_vtk_fk`, `_VALID_DOC_TYPES`), and the two route
handlers (`new_draft_page`, `create_draft_handler`) into `_upload.py`.

`__init__.py` shrinks from 2621 to 2038 lines (-583) and re-exports
every moved name via `__all__` so direct imports keep working without
a swap. The upload-touching tests in `tests/test_docs_routes.py`
(`TestCreateDraftHandler`, `TestUploadWithVtkLinking`,
`TestFlashMessages.test_upload_success_flashes_toast_on_detail`) had
their `patch("app.docs.routes.X")` paths rewritten to
`patch("app.docs.routes._upload.X")` for the moved targets
(`handle_upload`, `log_draft_upload`, `_connect`, `list_vtks_for_org`),
matching the documented post-#704 contract.

`tests/test_docs_routes_patch_paths.py` extends the PR-B pin tests with
three more assertions covering `_upload._validate_parent_vtk_fk`:
package-level patches don't reach `_upload`'s globals, the submodule
owns its module-load binding, and `patch("app.docs.routes._upload.X")`
DOES intercept the upload caller. Test count: 4 PR-B + 3 PR-C = 7,
all passing.

PR-D and PR-E will move the remaining handler bodies (`_detail`,
list/lifecycle helpers) and bring the package under the 400-line cap.

Verification:
- `uv run pytest tests/` — 2185 passed, 23 skipped (was 2178+ baseline)
- `uv run pytest tests/test_docs_routes_patch_paths.py` — 7 passed
- `uv run pytest tests/test_docs_routes_registered.py` — 1 passed
- `uv run ruff check + format` — clean
- `uv run pyright app/docs/routes/` — 0 errors
